separate testing::Error from crate::Error
