+++ libs/gesha-core/src/testing/error/mod.rs
@@ -1,0 +1,25 @@
+use std::path::PathBuf;
+use tokio::task::JoinError;
+
+#[derive(Debug)]
+pub enum Error {
+    DiffDetected {
+        output: String,
+        actual: PathBuf,
+        expected: PathBuf,
+    },
+    JoinError {
+        schema_path: PathBuf,
+        cause: JoinError,
+    },
+    ThreadNotFound(String),
+    UnknownTestCase {
+        path: String,
+    },
+}
+
+impl From<Error> for crate::Error {
+    fn from(e: Error) -> Self {
+        crate::Error::Testing(e)
+    }
+}
